fix(reflow): satisfy linter checks
