--- pkg/reflow/source/ebpf/conntrack_unsupported.go
@@ -1,5 +1,0 @@
-//go:build !linux || reflow_noebpf
-
-package ebpf
-
-type conntrackTracker struct{}
